Signed ints OP

+++ include/pid.h
@@ -19,7 +19,7 @@ public:
 	PID(int16_t* input, int16_t* output, int16_t* setpoint, uint16_t dt): input(input), output(output), setpoint(setpoint), prev_input(0), accumulator(0), dt(dt), min(0), max(0) {};
 	~PID() {};
 	void run();
-	void set_coef(uint32_t kp_, uint32_t ki_, uint32_t kd_) { kp = kp_; kis = ki_ * dt; kd = kd_; };
+	void set_coef(int32_t kp_, int32_t ki_, int32_t kd_) { kp = kp_; kis = ki_ * dt; kd = kd_; };
 	void set_limits(int16_t min_, int16_t max_) { min = min_; max = max_; };
 	uint32_t get_kp() { return kp; };
 	uint32_t get_ki() { return kis / dt; };
@@ -32,10 +32,10 @@ private:
 
 	int16_t prev_input;
 	int32_t accumulator;
-	uint32_t kp;
-	uint32_t kis;			// Scale then integrate
-	uint32_t kd;
-	uint16_t dt;
+	int32_t kp;
+	int32_t kis;			// Scale then integrate
+	int32_t kd;
+	int32_t dt;
 
 	int16_t min;
 	int16_t max;
